requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
owlready2==0.45
pyparsing==3.1.1
python-dotenv==1.0.0
rdflib==7.0.0
requests==2.31.0
six==1.16.0
urllib3==2.1.0
